Move Cards components into directories
