+++ assets/javascripts/kitten/components/cards/project-card/index.js
@@ -2,18 +2,18 @@ import React, { Component, Fragment } from 'react'
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
 import deprecated from 'prop-types-extra/lib/deprecated'
-import { card } from '../../hoc/card'
-import { Marger } from '../../components/layout/marger'
-import { Title } from '../../components/typography/title'
-import { Paragraph } from '../../components/typography/paragraph'
-import { Progress } from '../../components/meters/progress'
-import { ButtonImage } from '../../components/buttons/button-image'
-import { IconBadge } from '../../components/notifications/icon-badge'
-import { LockIcon } from '../../components/icons/lock-icon'
-import { CheckedCircleIcon } from '../../components/icons/checked-circle-icon'
-import { TagList } from '../../components/lists/tag-list'
-import { TypologyTagIcon } from '../../components/icons/typology-tag-icon'
-import { InstrumentTagIcon } from '../../components/icons/instrument-tag-icon'
+import { card } from '../../../hoc/card'
+import { Marger } from '../../../components/layout/marger'
+import { Title } from '../../../components/typography/title'
+import { Paragraph } from '../../../components/typography/paragraph'
+import { Progress } from '../../../components/meters/progress'
+import { ButtonImage } from '../../../components/buttons/button-image'
+import { IconBadge } from '../../../components/notifications/icon-badge'
+import { LockIcon } from '../../../components/icons/lock-icon'
+import { CheckedCircleIcon } from '../../../components/icons/checked-circle-icon'
+import { TagList } from '../../../components/lists/tag-list'
+import { TypologyTagIcon } from '../../../components/icons/typology-tag-icon'
+import { InstrumentTagIcon } from '../../../components/icons/instrument-tag-icon'
 
 export const MIN_WIDTH = 280
 export const MARGIN_BETWEEN = 40
