+++ assets/javascripts/kitten/components/cards/project-similar-card/index.js
@@ -1,17 +1,17 @@
 import React, { Component } from 'react'
 import classNames from 'classnames'
-import { card } from '../../hoc/card'
-import { Marger } from '../../components/layout/marger'
-import { Title } from '../../components/typography/title'
-import { Paragraph } from '../../components/typography/paragraph'
-import { ButtonIcon } from '../../components/buttons/button-icon'
-import { Separator } from '../../components/layout/separator'
-import { TagList } from '../../components/lists/tag-list'
-import { ArrowIcon } from '../../components/icons/arrow-icon'
-import { LockIcon } from '../../components/icons/lock-icon'
-import { Loader } from '../../components/loaders/loader'
-import { TypologyTagIcon } from '../../components/icons/typology-tag-icon'
-import { InstrumentTagIcon } from '../../components/icons/instrument-tag-icon'
+import { card } from '../../../hoc/card'
+import { Marger } from '../../../components/layout/marger'
+import { Title } from '../../../components/typography/title'
+import { Paragraph } from '../../../components/typography/paragraph'
+import { ButtonIcon } from '../../../components/buttons/button-icon'
+import { Separator } from '../../../components/layout/separator'
+import { TagList } from '../../../components/lists/tag-list'
+import { ArrowIcon } from '../../../components/icons/arrow-icon'
+import { LockIcon } from '../../../components/icons/lock-icon'
+import { Loader } from '../../../components/loaders/loader'
+import { TypologyTagIcon } from '../../../components/icons/typology-tag-icon'
+import { InstrumentTagIcon } from '../../../components/icons/instrument-tag-icon'
 
 class SimilarProjectCardComponent extends Component {
   constructor() {
