+++ assets/javascripts/kitten/components/cards/similar-projects-card/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { SimilarProjectCard } from '../../components/cards/project-similar-card'
+import { SimilarProjectCard } from '../../../components/cards/project-similar-card'
 
 export class SimilarProjectsCard extends Component {
   constructor() {
